[PM-3905] chore: move abort manager to vault
