+++ apps/browser/src/vault/background/abort-manager.ts
